requirements.txt
asyncpraw
praw
pyyaml
pandas
python-dotenv
openpyxl
google-play-scraper
app_store_scraper
boto3
botocore.exceptions

